% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{boxes}
\alias{boxes}
\title{boxes}
\usage{
boxes(points, prec)
}
\arguments{
\item{points}{: matrix of the objectives}

\item{prec}{: (double, length = nobj) desired accuracy for the objectives (edges of the boxes)}
}
\value{
vector of numbers for the boxes. boxes[i] gives the number of the box containing points[i].
}
\description{
This function returns a box number for each points individual of the population
}
\author{
Fabrice Zaoui
}
