% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caRamel.R
\name{caRamel}
\alias{caRamel}
\title{caRamel}
\usage{
caRamel(nobj, nvar, minmax, bounds, func, popsize, archsize, maxrun, prec,
  repart_gene = c(5, 5, 5, 5), gpp = NULL, blocks = NULL, pop = NULL,
  funcinit = NULL, noms_obj = NULL, listsave = NULL, write_gen = 0,
  carallel = TRUE, numcores = NULL)
}
\arguments{
\item{nobj}{: (integer, length = 1) the number of objectives to optimize (nobj >= 2)}

\item{nvar}{: (integer, length = 1) the number of variables}

\item{minmax}{: (logical, length = nobj) the objective is either a minimization (FALSE value) or a maximization (TRUE value)}

\item{bounds}{: (matrix, nrow = nvar, ncol = 2) lower and upper bounds for the variables}

\item{func}{: the name of the objective function to optimize. The function has to return 'nobj' values with : nobj >= 2}

\item{popsize}{: (integer, length = 1) the population size for the genetic algorithm}

\item{archsize}{: (integer, length = 1) the size of the Pareto front}

\item{maxrun}{: (integer, length = 1) the max. number of simulations allowed}

\item{prec}{: (double, length = nobj) the desired accuracy for the optimization of the objectives}

\item{repart_gene}{: (integer, length = 4) number of new solutions for each rule and per generation}

\item{gpp}{: (integer, length = 1) calling frequency for the rule "Fireworks"}

\item{blocks}{: groups for parameters}

\item{pop}{: (matrix, nrow = nset, ncol = nvar or nvar+nobj ) Initial population}

\item{funcinit}{: the name of the initialization function. The arguments are cl and numcores}

\item{noms_obj}{: the name of the objectives}

\item{listsave}{: names of the listing files. Default: None (no output)}

\item{write_gen}{: (integer, length = 1) if = 1, save files 'pmt' and 'obj' at each generation (= 0 by default)}

\item{carallel}{: (logical, length = 1) do parallel computations (TRUE by default)}

\item{numcores}{: (integer, length = 1) the number of cores for the parallel computations}
}
\value{
List of five elements:
\describe{
\item{success}{return value (logical, length = 1) : TRUE if successfull}
\item{parameters}{Pareto front (matrix, nrow = archsize, ncol = nvar)}
\item{objectives}{objectives of the Pareto front (matrix, nrow = archsize, ncol = nobj)}
\item{save_crit}{evolution of the maximum objectives}
\item{total_pop}{total population (matrix, nrow = popsize+archsize, ncol = nvar+nobj+ncomplement)}
}
}
\description{
R version of the multi-objective optimizer 'CaRaMEL' originally written for Scilab by Nicolas Le Moine
}
\details{
Documentation : "Principe de l'optimiseur CaRaMEL et illustration au travers d'exemples de parametres dans le cadre de la modelisation hydrologique conceptuelle"
                Frederic Hendrickx (EDF) and Nicolas Le Moine (UPMC)
                Report EDF H-P73-2014-09038-FR
}
\author{
Fabrice Zaoui - Celine Monteil
}
