% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cextrap.R
\name{Cextrap}
\alias{Cextrap}
\title{Cextrap}
\usage{
Cextrap(param, crit, directions, longu, n)
}
\arguments{
\item{param}{matrice [ NPoints , NPar ] des parametres deja evalues}

\item{crit}{matrice [ Npoints , NObj ] des criteres associes}

\item{directions}{matrice [ NDir, 2 ] les points de debut et de fin des vecteurs directeurs candidats}

\item{longu}{matrice [ NDir , 1 ] donnant la longueur de chaque segment ainsi defini dans l'espace des OBJ (mesure de la probabilite d'explorer cette direction)}

\item{n}{nombre de nouveaux vecteurs a generer}
}
\value{
xnouv matrice [ n , NPar ] des nouveaux vecteurs

pcrit : matrice [ n , NObj ] positions estimees des nouveaux jeux dans l'espace des objectifs
}
\description{
La fonction Cextrap propose n nouveaux candidats par extrapolation le long de directions orthogonales au Front de Pareto dans l'espace des objectifs
}
\author{
F. Zaoui
}
