% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cinterp.R
\name{Cinterp}
\alias{Cinterp}
\title{Cinterp}
\usage{
Cinterp(param, crit, simplices, volume, n)
}
\arguments{
\item{param}{: matrix [ NPoints , NPar ] of already evaluated parameters}

\item{crit}{: matrix [ Npoints , NObj ] of associated criteria}

\item{simplices}{: matrix [ NSimp , NObj+1 ] containing all or part of the triangulation of the space of the objectives}

\item{volume}{: matrix [ NSimp , 1 ] giving the volume of each simplex (measure of the probability of interpolating in this simplex)}

\item{n}{: number of new vectors to generate}
}
\value{
xnew : matrix [ n , NPar ] of new vectors

pcrit : matrix [ n , NObj ] estimated positions of new sets in the goal space
}
\description{
proposes n new candidates by interpolation in simplexes of the objective space
}
\author{
Fabrice Zaoui
}
