% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crecombination.R
\name{Crecombination}
\alias{Crecombination}
\title{Crecombination}
\usage{
Crecombination(param, blocs, n)
}
\arguments{
\item{param}{matrice [ . , NPar ] de la population des parametre}

\item{blocs}{liste de vecteurs d'entier : liste de blocs de variables pour la recombinaison}

\item{n}{nombre de nouveaux vecteurs a generer}
}
\value{
xnew matrice [ n , NPar ] des nouveaux vecteurs
}
\description{
La fonction Crecombination effectue une recombinaison des jeux de parametres.
}
\author{
F. Zaoui
}
