% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cusecovar.R
\name{Cusecovar}
\alias{Cusecovar}
\title{Cusecovar}
\usage{
Cusecovar(xref, amplif, n)
}
\arguments{
\item{xref}{matrice [ . , NPar ] de la population de reference dont on souhaite utiliser la structure de covariance}

\item{amplif}{facteur d'amplification de l'ecart-type sur chaque parametre}

\item{n}{nombre de nouveaux vecteurs a generer}
}
\value{
xnew matrice [ n , NPar ] des nouveaux vecteurs
}
\description{
La fonction Cusecovar Propose de nouveaux vecteurs de parametres respectant une structure de covariance.
}
\author{
F. Zaoui
}
