% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrease_pop.R
\name{decrease_pop}
\alias{decrease_pop}
\title{decrease_pop}
\usage{
decrease_pop(matobj, minmax, prec, archsize, popsize)
}
\arguments{
\item{matobj}{: matrix of objectives, dimension (ngames, nobj)}

\item{minmax}{: vector of booleans, of dimension nobj: TRUE if maximization of the objective, FALSE otherwise}

\item{prec}{: nobj dimension vector: accuracy}

\item{archsize}{: integer: archive size}

\item{popsize}{: integer: population size}
}
\value{
A list containing two elements:
\describe{
\item{ind_arch}{}
\item{ind_pop}{} 
}
}
\description{
decreases the population of parameter sets
}
\author{
Fabrice Zaoui
}
