% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dimprove.R
\name{Dimprove}
\alias{Dimprove}
\title{Dimprove}
\usage{
Dimprove(o_splx, f_splx)
}
\arguments{
\item{o_splx}{: matrix of objectives of simplexes (nrow = npoints, ncol = nobj)}

\item{f_splx}{: vector (npoints) of associated Pareto numbers (1 = dominated)}
}
\value{
list of elements "oriedge": oriented edges and "ledge": length
}
\description{
determines directions for improvement
}
\author{
Fabrice Zaoui
}
