% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominate.R
\name{dominate}
\alias{dominate}
\title{dominate}
\usage{
dominate(matobj)
}
\arguments{
\item{matobj}{matrice [ NInd , NObj ] des objectifs}
}
\value{
f vecteur de dimension NInd des dominances
}
\description{
La fonction dominate calcule les fronts de Pareto successifs d'une population (classement "en pelure d'oignon")
}
\author{
F. Zaoui
}
