% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominated.R
\name{dominated}
\alias{dominated}
\title{dominated}
\usage{
dominated(x, Y)
}
\arguments{
\item{x}{: vecteur ligne}

\item{Y}{: matrice}
}
\value{
D : vecteur de booleens
}
\description{
La fonction dominated indique quelles lignes de la matrice Y sont dominees par le vecteur (ligne) x
}
\author{
F. Zaoui
}
