% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matvcov.R
\name{matvcov}
\alias{matvcov}
\title{matvcov}
\usage{
matvcov(x, g)
}
\arguments{
\item{x}{: population}

\item{g}{: center of reference population (in the parameter space)}
}
\value{
rr : variances-covariances matrix on the reference population
}
\description{
calculates the variances-covariances matrix on the reference population
}
\author{
Fabrice Zaoui
}
