% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto.R
\name{pareto}
\alias{pareto}
\title{pareto}
\usage{
pareto(X)
}
\arguments{
\item{X}{matrice [NInd * NObj]}
}
\value{
Ft matrice colonne [NInd * 1]
}
\description{
La fonction pareto indique quelles lignes de la matrice de criteres X sont Pareto
}
\author{
F. Zaoui
}
