% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rselect.R
\name{rselect}
\alias{rselect}
\title{rselect}
\usage{
rselect(n, facp)
}
\arguments{
\item{n}{: nombre de points a selectionner}

\item{facp}{: vecteur de points initiaux}
}
\value{
ix : rangs des points selectionnes (vecteur de dimension n)
}
\description{
La fonction rselect effectue une selection de n points dans facp
}
\author{
F. Zaoui
}
