% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol_splx.R
\name{vol_splx}
\alias{vol_splx}
\title{vol_splx}
\usage{
vol_splx(S)
}
\arguments{
\item{S}{: matrice (d+1) lignes * d colonnes contenant les coordonnees en d-dim des d+1 sommets d'un simplex}
}
\value{
V volume du simplex
}
\description{
La fonction vol_splx calcule le volume d'un simplex
}
\author{
F. Zaoui
}
