% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{boxes}
\alias{boxes}
\title{boxes}
\usage{
boxes(points, prec)
}
\arguments{
\item{points}{: matrice des objectifs}

\item{prec}{(double, length = nobj) precision souhaitee pour les objectifs (arretes des boites)}
}
\value{
vecteur des numeros de boites
}
\description{
La fonction boxes renvoie un numero de boite par points d'une population
}
\author{
F. Zaoui
}
