% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{boxes}
\alias{boxes}
\title{Box numbering for each points individual of the population}
\usage{
boxes(points, prec)
}
\arguments{
\item{points}{: matrix of the objectives}

\item{prec}{: (double, length = nobj) desired accuracy for the objectives (edges of the boxes)}
}
\value{
vector of numbers for the boxes. boxes[i] gives the number of the box containing points[i].
}
\description{
This function returns a box number for each points individual of the population
}
\examples{
# Definition of the parameters
points <- matrix(rexp(200), 100, 2)
prec <- c(1.e-3, 1.e-3)
# Call the function
res <- boxes(points, prec)

}
\author{
Fabrice Zaoui
}
