% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caRamel.R
\name{caRamel}
\alias{caRamel}
\title{caRamel}
\usage{
caRamel(nobj, nvar, minmax, bounds, func, popsize, archsize, maxrun, prec,
  repart_gene = c(5, 5, 5, 5), gpp = NULL, blocs = NULL, pop = NULL,
  funcinit = NULL, noms_obj = NULL, listsave = NULL, ecrit_gen = 0,
  parallele = TRUE, numcores = NULL)
}
\arguments{
\item{nobj}{(integer, length = 1) the number of objectives to optimize (nobj >= 2)}

\item{nvar}{(integer, length = 1) the number of variables}

\item{minmax}{(logical, length = nobj) the objective is either a minimization (FALSE value) or a maximization (TRUE value)}

\item{bounds}{(matrix, nrow = nvar, ncol = 2) lower and upper bounds for the variables}

\item{func}{the name of the objective function to optimize. The function has to return 'nobj' values with : nobj >= 2}

\item{popsize}{(integer, length = 1) the population size for the genetic algorithm}

\item{archsize}{(integer, length = 1) the size of the Pareto front}

\item{maxrun}{(integer, length = 1) the max. number of simulations allowed}

\item{prec}{(double, length = nobj) the desired precision for the optimization of the objectives}

\item{repart_gene}{(integer, length = 4) nombre de jeux generes pour chaque regle par generation,}

\item{gpp}{(integer, length = 1) frequence d'appel de la regle "Fireworks"}

\item{blocs}{blocs pour les parametres}

\item{pop}{(matrix, nrow = njeux, ncol = nvar ou nvar+nobj ) Population initiale}

\item{funcinit}{the name of the initialization function. The arguments are is cl and numcores.}

\item{noms_obj}{the name of the objectives}

\item{listsave}{liste du noms des fichiers de suivi ecrits en cours d'optim, pas d'ecriture par defaut. Champs obligatoires : "pmt" (fichier des jeux de parametres sur le front), "obj" (Objectifs associes), "evol" (evolution des objectifs max par generation). Champ optionnel : "totalpop" (population totale : parametres et objectifs associes)}

\item{ecrit_gen}{(integer, length = 1) si =1, conservation des fichiers pmt et obj a chaque generation (0 par defaut)}

\item{parallele}{(logical, length = 1) activation du calcul parallele (TRUE par defaut)}

\item{numcores}{(integer, length = 1) the number of cores for parallel computation}
}
\value{
List of five elements:
\describe{
\item{success}{return value (logical, length = 1) : TRUE if successfull}
\item{parameters}{Pareto front (matrix, nrow = archsize, ncol = nvar)}
\item{objectives}{objectives of the Pareto front (matrix, nrow = archsize, ncol = nobj)}
\item{suivi_crit}{evolution of the maximum objectives}
\item{pop_totale}{total population (matrix, nrow = popsize+archsize, ncol = nvar+nobj+ncomplement)}
}
}
\description{
R version of the multi-objective optimizer 'CaRaMEL' originally written for Scilab by Nicolas Le Moine
}
\details{
Documentation : "Principe de l'optimiseur CaRaMEL et illustration au travers d'exemples de parametres dans le cadre de la modelisation hydrologique conceptuelle"
                Frederic Hendrickx (EDF) and Nicolas Le Moine (UPMC)
                Report EDF H-P73-2014-09038-FR
}
\author{
Fabrice Zaoui - Celine Monteil
}
