% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cextrap.R
\name{Cextrap}
\alias{Cextrap}
\title{Cextrap}
\usage{
Cextrap(param, crit, directions, longu, n)
}
\arguments{
\item{param}{: matrix [ NPoints , NPar ] of already evaluated parameters}

\item{crit}{: matrix [ Npoints , NObj ] of associated criteria}

\item{directions}{: matrix [ NDir, 2 ] the starting and ending points of the candidate vectors}

\item{longu}{: matrix [ NDir , 1 ] giving the length of each segment thus defined in the OBJ space (measure of the probability of exploring this direction)}

\item{n}{: number of new vectors to generate}
}
\value{
xnouv : matrix [ n , NPar ] of new vectors

pcrit : matrix [ n , NObj ] estimated positions of new sets in the goal space
}
\description{
gives n new candidates by extrapolation along orthogonal directions to the Pareto front in the space of the objectives
}
\author{
Fabrice Zaoui
}
