% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cinterp.R
\name{Cinterp}
\alias{Cinterp}
\title{Cinterp}
\usage{
Cinterp(param, crit, simplices, volume, n)
}
\arguments{
\item{param}{matrice [ NPoints , NPar ] des parametres deja evalues}

\item{crit}{matrice [ Npoints , NObj ] des criteres associes}

\item{simplices}{matrice [ NSimp , NObj+1 ] contenant tout ou partie de la triangulation de l'espace des objectifs}

\item{volume}{matrice [ NSimp , 1 ] donnant le volume de chaque simplexe (mesure de la probabilite d'interpoler dans ce simplexe)}

\item{n}{nombre de nouveaux vecteurs a generer}
}
\value{
xnouv matrice [ n , NPar ] des nouveaux vecteurs

pcrit : matrice [ n , NObj ] positions estimees des nouveaux jeux dans l'espace des objectifs
}
\description{
La fonction Cinterp propose n nouveaux candidats par interpolation dans des simplexes de l'espace des objectifs.
}
\author{
F. Zaoui
}
