% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crecombination.R
\name{Crecombination}
\alias{Crecombination}
\title{Crecombination}
\usage{
Crecombination(param, blocks, n)
}
\arguments{
\item{param}{: matrix [ . , NPar ] of the population of parameters}

\item{blocks}{: list of integer vectors: list of variable blocks for recombination}

\item{n}{: number of new vectors to generate}
}
\value{
xnew : matrix [ n , NPar ] of new vectors
}
\description{
performs a recombination of the sets of parameters
}
\author{
Fabrice Zaoui
}
