% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrease_pop.R
\name{decrease_pop}
\alias{decrease_pop}
\title{decrease_pop}
\usage{
decrease_pop(matobj, minmax, prec, archsize, popsize)
}
\arguments{
\item{matobj}{matrice des objectifs, de dimension (njeux, nobj)}

\item{minmax}{: vecteur de booleens, de dimension nobj : TRUE si maximisation de l'objectif, FALSE sinon}

\item{prec}{: vecteur de dimension nobj : precision}

\item{archsize}{: entier, taille de l'archive}

\item{popsize}{: entier : taille de la population}
}
\value{
Une liste contenant deux elements:
\describe{
\item{ind_arch}{}
\item{ind_pop}{} 
}
}
\description{
La fonction decrease_pop reduit la population de jeux de parametres.
}
\author{
F. Zaoui
}
