% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dimprove.R
\name{Dimprove}
\alias{Dimprove}
\title{Dimprove}
\usage{
Dimprove(o_splx, f_splx)
}
\arguments{
\item{o_splx}{matrices des objectifs des simplexes (nrow = npoints, ncol = nobj)}

\item{f_splx}{vecteur (npoints) des numeros de Pareto associes (1 = domine)}
}
\value{
liste des elements "oriedge" : aretes orientees et "ledge" : longueur
}
\description{
La fonction Dimprove determine des directions d'amelioration
}
\author{
F. Zaoui
}
