% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dimprove.R
\name{Dimprove}
\alias{Dimprove}
\title{Determination of directions for improvement}
\usage{
Dimprove(o_splx, f_splx)
}
\arguments{
\item{o_splx}{: matrix of objectives of simplexes (nrow = npoints, ncol = nobj)}

\item{f_splx}{: vector (npoints) of associated Pareto numbers (1 = dominated)}
}
\value{
list of elements "oriedge": oriented edges and "ledge": length
}
\description{
determines directions for improvement
}
\examples{
# Definition of the parameters
o_splx <- matrix(rexp(6), 3, 2)
f_splx <- c(1,1,1)
# Call the function
res <- Dimprove(o_splx, f_splx)

}
\author{
Fabrice Zaoui
}
