% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominate.R
\name{dominate}
\alias{dominate}
\title{dominate}
\usage{
dominate(matobj)
}
\arguments{
\item{matobj}{: matrix [ NInd , NObj ] of objectives}
}
\value{
f : vector of dimension NInd of dominances
}
\description{
calculates the successive Pareto fronts of a population (classification "onion peel")
}
\examples{
# Definition of the parameters
matobj <- matrix(rexp(200), 100, 2)
# Call the function
res <- dominate(matobj)

}
\author{
Fabrice Zaoui
}
