% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominated.R
\name{dominated}
\alias{dominated}
\title{dominated}
\usage{
dominated(x, Y)
}
\arguments{
\item{x}{: row vecteur}

\item{Y}{: matrix}
}
\value{
D : vector of booleans
}
\description{
indicates which rows of the matrix Y are dominated by the vector (row) x
}
\author{
F. Zaoui
}
