% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsize.R
\name{downsize}
\alias{downsize}
\title{downsize}
\usage{
downsize(points, Fo, prec)
}
\arguments{
\item{points}{: matrix of objectives}

\item{Fo}{: rank on the front of each point (1: dominates on the Pareto)}

\item{prec}{: (double, length = nobj) desired accuracy for sorting objectives}
}
\value{
vector indices
}
\description{
La fonction downsize renvoie les indentifiants d'une population en ne conservant qu'un seul point par boite (de resolution prec)
}
\author{
Fabrice Zaoui
}
