% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsize.R
\name{downsize}
\alias{downsize}
\title{downsize}
\usage{
downsize(points, Fo, prec)
}
\arguments{
\item{points}{: matrice des objectifs}

\item{Fo}{: rang sur le front de chaque point (1 : domine sur le Pareto)}

\item{prec}{(double, length = nobj) precision souhaitee pour le tri des objectifs}
}
\value{
vecteur des indices
}
\description{
La fonction downsize renvoie les indentifiants d'une population en ne conservant qu'un seul point par boite (de resolution prec)
}
\author{
F. Zaoui
}
