% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matvcov.R
\name{matvcov}
\alias{matvcov}
\title{matvcov}
\usage{
matvcov(x, g)
}
\arguments{
\item{x}{population}

\item{g}{barycentre de la population de reference (dans l'espace des parametres)}
}
\value{
rr matrice de variances-covariances sur la population de reference
}
\description{
La fonction matvcov calcule la matrice de variances-covariances sur la population de reference
}
\author{
F. Zaoui
}
