% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newXval.R
\name{newXval}
\alias{newXval}
\title{newXval}
\usage{
newXval(param, crit, isperf, sp, bounds, repart_gene, blocs, fireworks)
}
\arguments{
\item{param}{matrice [ Nvec , NPar ] des parametres de la population courante}

\item{crit}{matrice [ Nvec , NObj ] des criteres associes}

\item{isperf}{vecteur de booleens de longueur NObj, TRUE si maximisation de l'objectif, FALSE sinon}

\item{sp}{Variance a priori des parametres}

\item{bounds}{bornes inf et sup des parametres [ NPar , 2 ]}

\item{repart_gene}{: matrice de longueur 4 donnant le nombre de jeux a generer avec chaque regle : 1 Interpolation dans les simplexes du front, 2 Extrapolation selon les directions des aretes "orthogonales" au front, 3 Tirages aleatoires avec matrice de variance-covariance prescrite, 4 Recombinaison par blocs fonctionnels}

\item{blocs}{liste de vecteurs d'entier contenant les blocs fonctionnels de parametres}

\item{fireworks}{: booleen, TRUE si on teste une variation aleatoire sur chaque parametre et chaque maximum de F.O.}
}
\value{
xnouv matrice de nouveaux vecteurs [ sum(Repart_Gene) + eventuellement (nobj+1)*nvar si fireworks , NPar ]

project_crit position supposee des nouveau vecteurs dans l'espace des criteres : [ sum(Repart_Gene)+ eventuellement (nobj+1)*nvar si fireworks , NObj ];
}
\description{
La fonction newXval genere une nouvelle population de jeux de parametres en suivant les cinq regles de caRamel.
}
\author{
F. Zaoui
}
