% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto.R
\name{pareto}
\alias{pareto}
\title{Indicates which rows are Pareto}
\usage{
pareto(X)
}
\arguments{
\item{X}{: matrix of objectives [NInd * NObj]}
}
\value{
Ft : vector [NInd], TRUE when the set is on the Pareto front.
}
\description{
indicates which rows of the X criterion matrix are Pareto, when objectives need to be maximized
}
\examples{
# Definition of the parameters
X <- matrix(runif(200), 100, 2)
# Call the function
is_pareto <- pareto(X)

}
\author{
Alban de Lavenne, Fabrice Zaoui
}
