% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto.R
\name{pareto}
\alias{pareto}
\title{pareto}
\usage{
pareto(X)
}
\arguments{
\item{X}{: matrix [NInd * NObj]}
}
\value{
Ft : column matrix [NInd * 1]
}
\description{
indicates which rows of the X criterion matrix are Pareto
}
\examples{
# Definition of the parameters
X <- matrix(rexp(200), 100, 2)
# Call the function
res <- dominate(X)

}
\author{
Fabrice Zaoui
}
