% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pareto.R
\name{plot_pareto}
\alias{plot_pareto}
\title{Plotting of a population of objectives and Pareto front}
\usage{
plot_pareto(MatObj, nobj = NULL, objnames = NULL)
}
\arguments{
\item{MatObj}{: matrix of the objectives [NInd, nobj]}
}
\description{
Plots graphs the population regarding each couple of objectives and emphasizes the Pareto front
}
\examples{
# Definition of the population
Pop <- matrix(runif(300), 100, 3)
# Call the function
plot_pareto(MatObj = Pop)

}
\author{
C. Monteil
}
