% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rselect.R
\name{rselect}
\alias{rselect}
\title{Selection of n points}
\usage{
rselect(n, facp)
}
\arguments{
\item{n}{: number of points to select}

\item{facp}{: vector of initial points}
}
\value{
ix : ranks of selected points (vector of dimension n)
}
\description{
performs a selection of n points in facp
}
\examples{
# Definition of the parameters
n <- 5
facp <- runif(30)
# Call the function
res <- rselect(n, facp)

}
\author{
Fabrice Zaoui
}
