% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val2rank.R
\name{val2rank}
\alias{val2rank}
\title{val2rank}
\usage{
val2rank(X, opt)
}
\arguments{
\item{X}{vecteur a traiter}

\item{opt}{entier qui donne la regle a suivre en cas  de rangs lies (valeurs repetees) : si opt = 1, on renvoie le rang moyen, si opt = 2, on renvoie le rang correspondant dans la serie des valeurs uniques, si opt = 3, on renvoie le rang max}
}
\value{
R vecteur des rangs
}
\description{
La fonction val2rank convertit les valeurs d'un vecteur en leur rang
}
\author{
F. Zaoui
}
