% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol_splx.R
\name{vol_splx}
\alias{vol_splx}
\title{vol_splx}
\usage{
vol_splx(S)
}
\arguments{
\item{S}{: matrix (d+1) rows * d columns containing the coordinates in d-dim of d + 1 vertices of a simplex}
}
\value{
V : simplex volume
}
\description{
calculates the volume of a simplex
}
\examples{
# Definition of the parameters
S <- matrix(rexp(6), 3, 2)
# Call the function
res <- vol_splx(S)

}
\author{
Fabrice Zaoui
}
